pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nfs-ml-suite"
version = "0.1.0"
description = "Near-field scan ML suite (commercial delivery layout) — forward/inverse training, predict, export."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
  "matplotlib>=3.8",
  "torch>=2.1",
  "tqdm>=4.66",
]

[project.scripts]
nfs-ml = "nfs_ml.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
